cli/session-manage: Make --resume and --restart run in background
Both the --resume and --restart modes of the openvpn3 session-manage
command would run in the foreground, which would result in stopping the
session if SIGINT/CTRL-C happened.  This is often not the intension.

Instead, let --resume and --restart modes run in the background by
default.  To revert to the old behaviour, add --timeout to the command
line with a high enough value.

Signed-off-by: David Sommerseth <davids@openvpn.net>