log: Ensure extracted systemd-journald fields have correct length
The log data extracted via sd_journal_get_data() did not restrict the
string extraction to the string length value returned by that function.
This resulted in garbage being added on some log entries.

When creating a string object of the journald log data, ensure this
length indicator is respected.

Signed-off-by: David Sommerseth <davids@openvpn.net>