client: Implement support for CR_TEXT pending authentication
If the server sends an AUTH_PENDING request, that is an indication of a
slightly different authentication process in the client side.  This
event also includes a timeout how long the server will accept waiting
for the client to provide further credentials.

A following INFO event will be sent by the server, which contains a
CR_TEXT indicator.  This event will contain the details to ask the
end-user.

When the client responds, a Control Channel message is sent via the
new SendAuthPendingReponse() method, the server will pick that response
up instantly; no further processing is required on the client side.

Signed-off-by: David Sommerseth <davids@openvpn.net>