selinux: Allow openvpn3_client_t to use syslog
Occasionally there can be seen SELinux denials when
openvpn3-service-client starts up, related to logging to syslog.

This seems to happen during the process start up where it will write a
few lines to stdout.  This is picked up by the dbus-daemon and sent to
the system log, but at this point the openvpn3-service-client process is
already confined as openvpn3_client_t which did not have this privilege.

Signed-off-by: David Sommerseth <davids@openvpn.net>