ovpn3cli: User credential input error handling input with session-start
The error message the user got if the user input was not sufficient or
the user aborted had lots of room for improvements.

This implements a better SIGINT handling in addition to catch errors
specific to the input from the user.  It will now also retry to get new
input if it was not initially accepted.

Signed-off-by: David Sommerseth <davids@openvpn.net>