netcfg: Exit early if systemd-resolved cannot be reached
This adds an early check if the org.freedesktop.resolve1 service can be
reached.  This ensures the net.openvpn.v3.netcfg service can report a
more understandable error message as early as possible.

When this error happens, it will attempt to log an error about the
unavailable systemd-resolved service and it will log that no DNS
resolver has been configured.

Signed-off-by: David Sommerseth <davids@openvpn.net>