dbus: Provide a generic ProxyConnect() method
For use cases where the DBusProxy is told to wait connecting to the
D-Bus service, the setup process would require two additional
SetupProxy() calls by the caller.  In many cases, this connection setup
will be the same as the one in the constructor, just delayed.

By providing a ProxyConnect() method, the proxy will be connected
properly - but in a single call.  And this method can also be reused
in the constructors as well, avoiding code duplication.

Signed-off-by: David Sommerseth <davids@openvpn.net>