dbus: Additional DBusProxy cleanups
Move over to member initialization while declaring it instead of in each
constructor.  A couple of class members were also declared at the wrong
location, which also removed a superfluous 'protected' statement.

The config.h should also be included, as we depend on a few macros
declared in this file.

Signed-off-by: David Sommerseth <davids@openvpn.net>