log: Strip double {tag:...} references in openvpn3-admin journal output
Depending on how the logging is configured, the log message can also
contain a logtag reference.  This ends up being duplicated in the output
of 'openvpn3-admin journal'.

When extracting the log message, look for the same logtag string in the
log message and remove it if it is found.

Signed-off-by: David Sommerseth <davids@openvpn.net>