configmgr: Rework the config profile usage counter check logic
The logic used to consider if the Fetch() method was from a backend VPN
client process was very simplistic and poor.  It only checked if the PID
value of the sender was from the 'openvpn' user account.

This reimplements this check by checking the unique bus ID from the
sender variable provided via callback_method_call() against a separate
lookup based on a recomposed well-known bus name and retrieving the
unique bus name for this well-known bus name.  These should match if the
caller is truly from the openvpn3-service-client process.

Signed-off-by: David Sommerseth <davids@openvpn.net>