dbus/glib: Extend GVariantBuilder helpers with type override
D-Bus differentiates between string and path types, while they are both
handled as strings in C/C++.  The type deduction logic in
GVariantBuilderAdd() and GVariantBuilderFromVector() will therefore not
work well when D-Bus expects the data type to object paths.

This extends these methods with an optional override_type variable which
can be used to specify a specific D-Bus type instead of solely relying on
the C++ type deduction logic in these cases.

Signed-off-by: David Sommerseth <davids@openvpn.net>