ovpn3cli: Improve configs-list output
The output of openvpn3 configs-list can be quite verbose which might
provide details not needed most of the time.  This implements a
--verbose mode which provides the details found in prior releases and
makes the default mode use just a single line of output per
configuration profile.

The date formatting has also been adjusted to use the ISO-8601
formatting (%F %X aka YYYY-mm-dd HH:MM:SS).

In the verbose mode, assigned tags to profiles will also be part
of the listing.

Signed-off-by: David Sommerseth <davids@openvpn.net>