client: Do not reset empty env array in backendstart
This is a minor issue on Linux. However it should be fixed as passing a
nullptr to execve() in the arg or env fields is generally not a good
idea.

The intention was to avoid passing any data to execve() unless strictly
needed.  But the env char * array is already empty from the start, so
there is no need to nullify it if it is not required.  This should be
more to the expectations.  The arg char * array cannot be empty, so no
concerns here.

This issue was reported by Coverity.

Signed-off-by: David Sommerseth <davids@openvpn.net>