dbus: Clean up odd init booleans in DBusProxy
These additional boolean values indicating which proxy object being
setup is not really helpful.  Instead ensure that the proxy pointers are
initialized correctly to nullptr in the start, and use that for checking
the state.

Also add an extra check that the proxy pointer is a glib2 based object
before deleting them.

Signed-off-by: David Sommerseth <davids@openvpn.net>