dbus: Cleaning up impropoer DBusProxy internal proxy pointers freeing
Several methods would re-use the 'proxy' variable locally by calling
SetupProxy() to a local GDBusProxy pointer.  This is sloppy and can
cause confusion with the global object variable with the same name.

Often these methods also did not unreference the local proxy object
when the method was done and returned.

Signed-off-by: David Sommerseth <davids@openvpn.net>