dbus: Remove the DBusProxy() constructor accepting DBus() objects
Even though passing of DBus objects to the DBusProxy constructor can
make sense and make the code clean, this approach turns out to cause
issues with also calling the destructor of the DBus object when the
DBusProxy object is destroyed.  This adds lots of other potential issues
along the way.

Since the whole D-Bus implementation is going to go through a massive
overhaul, just pass the GDBusConnection pointer instead for now and
don't close the connection itself inside the DBusProxy object itself.

Signed-off-by: David Sommerseth <davids@openvpn.net>