configmgr: Fix incorrect set_override() declaration
The set_override() method is called with std::string as the key data
type.  When this method is set up to accept gchar *, the compiler would
take the template version instead with and passing the GVariant pointer
to a method not capable of extracting any value from this.  This
resulted in empty override settings.

Signed-off-by: David Sommerseth <davids@openvpn.net>