core/config: Extend OpenVPN Access Server configuration support
In commit 6993bd91859bc4b10 some OpenVPN Access Server specific changes
was added to avoid issues related to meta-options used in configuration
files from this product.

This pulls in a lot more meta-options which has been found in OpenVPN
Access Server.  Not all of these is in use in modern Access Server
configurations nor are all of them relevant or possible in OpenVPN 3
Linux.  But they are added to avoid hitting corner cases with some
older configurations.

Signed-off-by: David Sommerseth <davids@openvpn.net>