common: Extend ParsedArgs::Present(std::vector<>) to optionally return empty string
By default the ParsedArgs::Present(std::vector<std::string>) method
will throw OptionNotFound if one of the listed options are not found.

In some cases, this will not work well. This extends this method to have
a no_throw bool flag which make this method just return an empty string
if there are no match.

Signed-off-by: David Sommerseth <davids@openvpn.net>