python: Ensure --profile-overrides is not put into generated config
The ConfigParser.GenerateConfig() handles the --profile-override as a
valid OpenVPN configuration file option.  This option is specific to the
OpenVPN 3 Python module to be able to set a profile override from the
command line.  These overrides should be set separately after the
generated configuration has been imported into the OpenVPN 3
Configuration Manager.

If this option is passed into the generated configuration, the OpenVPN 3
client backend will complain and bail out when it sees unknown
configuration options.  This is a new behaviour after OpenVPN 3 Linux
updated to use OpenVPN 3 Core library v3.8.

Signed-off-by: David Sommerseth <davids@openvpn.net>