ovpn3cli/config: Extend JSON schema used by config-dump --json
The value of this JSON formatted dump is limited, as there are more
information available which is useful.  This change will add almost all
the JSON fields found in the JSON schema used by the Configuration
Manager when storing persistent configurations to disk.

This can make it possible to export a configuration to be imported
another place.

Signed-off-by: David Sommerseth <davids@openvpn.net>