configmgr/proxy: Add forcing feature loading in the OpenVPN3ConfigurationProxy
The OpenVPN3ConfigurationProxy tries to not do too much in the
constructor phase to be quick in the setup step.  But in some cases
the feature list should be loaded as early as possible.  By setting
the optional 'force_feature_load' flag, the available backend
features will be queried and set up accordingly.

Signed-off-by: David Sommerseth <davids@openvpn.net>