proxy: Fix incorrect version extraction logic with non-service root paths
When trying to extract the service version when the requested proxy
object is pointing on a child object, the GetServiceVersion() method
would fail unless that child object contained the 'version' property -
which again could be a wrong property value.

Rework the whole logic to use the D-Bus service root path to extract the
version.  And do an internal caching value of this version string to
avoid more D-Bus calls than strictly needed.

Signed-off-by: David Sommerseth <davids@openvpn.net>