build: Rework the OpenVPN 3 Core library version macro (OPENVPN_VERSION)
Until now, the OPENVPN_VERSION macro has been added to the compiler
command line via the CXXFLAGS env variable.  While this works, it
requires all Makefiles to be rebuilt when updating the Core library
during development.

This puts the OPENVPN_VERSION macro as a define into config.h or
config-version.h.  Which file depends on if the source code is a git
checkout or not.  To get the right OpenVPN 3 Core library version, only
the config-version.h need to be rebuilt as well as all the sources
including this header.  For release builds based on the tarball, the
version will be copied from the openvpn3-core-version file into
config.h.

Signed-off-by: David Sommerseth <davids@openvpn.net>