ovpn3cli/admin: log-service lacked initial state and init
The log-level and boolean flags where the LogServiceProxy object tracks
changes was not properly set to an initial state.  In addition the
variable tracking the changes was not properly initialized.

This adds a new Changed:INITIALIZED flag, which is used to track if the
current initial settings has been recorded.  If this is not set, that
information will be cached once before doing any changes and the
Changed:INITIALIZED is set.  This gives a better chance to spot the
changes properly.

Signed-off-by: David Sommerseth <davids@openvpn.net>