GDBus++: First step in migrating to the new D-Bus implementation
This step takes all the src/common code and some related parts and
builds it as a locally static libcommon.a library.  This will be used
during linking of the other binaries.

As part of this first step, the test programs for the RequiresQueue API
was also migrated and is buildable.

During this migration, no longer used code has been removed and there
has been some API cleanup as well.

Signed-off-by: David Sommerseth <davids@openvpn.net>