build: Add bare meson setup for generating man pages
This adds the tri-state generate-man configuration option (auto,
enabled, disabled).  The default is auto.

If rst2man or rst2man.py is not found, man pages will not be generated
if the state is set to 'auto'.  If the state is 'enabled', then the
build configuration/setup will fail if rst2man or rst2man.py cannot be
located.

Signed-off-by: David Sommerseth <davids@openvpn.net>