build: Build and install SELinux policies via Meson
The approach here is a bit comprehensive than what it ideally
could be.  But I did not find it worth the effort to dissect
the SELinux policy compilation and package building even
further.

The building itself is trivial using the Makefile provided
by the SELinux policy development chain.  The tricky detail
is that an SELinux policy isn't a single source file.  Each
policy file share the same "basename" with different
extensions. In addition, the output should ideally be bzip2
compressed.  On top of that, the SELinux Makefile is not
really designed for out-of-tree builds.

These details makes it harder to make Meson just run 'make'.
And that is the purpose of the scripts/build-selinux-policy
helper script, to make it easier for Meson to just to kick
off a job to get the policy package compiled and pick up
the resulting compressed package.

Signed-off-by: David Sommerseth <davids@openvpn.net>