build: Prepare the ground for building dco-keyconfig.proto
The dco-keyconfig.proto part creates the protobuf interface as C++
source, used to communicate with the ovpn-dco kernel.

This was located under src/netcfg but has been moved to src/dco,
as this is shared between more components.

All the build targets needed to build the C++ source and header files
are now handled in its own meson.build file as well for simplicity.

Signed-off-by: David Sommerseth <davids@openvpn.net>