GDBus++: Migrate the unit tests
Almost all of the unit-tests (with one exception) is now ready
to be migrated to GDBus++ and the Meson build system.

The tests dns-settings-manager-test.cpp is still depending on more of
the NetCfg infrastructure to be migrated.  For now, just exclude these
tests from the unit test.  It can easily be re-enabled once
NetCfg::DNS::SettingsManager and related components has been migrated.

Since the unit tests is based on the googletest framework, make use of
the Meson wrap feature to easily get the needed gtest dependency.

Signed-off-by: David Sommerseth <davids@openvpn.net>