client/log: Rewrite the Core library D-Bus log implementation
The current implementation of the CoreDBusLogBase used the Core
library's openvpn/log/logbase.hpp and openvpn/log/logthread.hpp
implementation.  That turned out to be very sensitive to #include
file ordering and did not fit that well into the current GDBus++
refactoring.  The end result was that Core debug/log events did
not appear anywhere.

Since only three OPENVPN_LOG* macros need to be defined plus
 support context struct (which can even be empty in this case),
replace the CoreDBusLogBase implementation with a simpler one.

The base logging principles of CoreDBusLogBase is now moved into
an internal CoreLog::DBusLogger class which is not needed to interact
with directly.  This class will take care sending the proper log
signals via a D-Bus connection.

In the core-dbus-logger.hpp, the publicly exposed API is declared.
This also sets up the needed OPENVPN_LOG*() macros and ensures the Core
library is able to send log events some where.  If the D-Bus layer has
not been initialized, it will default to use std::cout instead.

Signed-off-by: David Sommerseth <davids@openvpn.net>