client/core: Reorder and fix #include file related challenges
To properly define certain features and implementations, the #include
file order is unfortunately a important when including header files for
the OpenVPN 3 Core Library.

The openvpn/io/io.hpp header has been included early, to ensure
openvpn_io is properly declared before including other Core library
headers.

The build-config.h should ideally not be included in a header file, but
still this seems to be the best way as there are functionality in these
core client implementation headers depending on some of the macros
coming from that header file.

The platform.h include is just relocated closer to the rest of the
Core library headers.

Signed-off-by: David Sommerseth <davids@openvpn.net>