tests/netcfg: List D-Bus paths if the "Re-fetching" fails
The whole check with "Re-fetching device paths" test is fairly fragile,
so for now we just list the devices it retrieved to manually verifying
the result.

This is easily triggered when running more netcfg-proxy-unit tests in
parallel.  This is not an error in the openvpn3-service-netcfg, but that
the this test lacks proper evaluation of the still existing paths.  This
will be resolved later on.

Signed-off-by: David Sommerseth <davids@openvpn.net>