netcfg/dns: Refactor NetCfg::DNS::SystemdResolved
Refactor to replace RC/RCPtr with std::shared_ptr and related
changes in the ResolverBackendInterface API.  With this
refactoring issues related to the device destruction surfaced
which resulted in segfault crashes when openvpn3-service-netcfg
was stress-tested.  The reason was that several objects was
removed in one thread while another thread was performing work
on them.

This was resolved by using std::shared_ptr for the updateQueueEntry
objects as well, to ensure these objects has a proper reference
counting before being removed from memory.

Another aspect was also that the previous inheritance model, where
SystemdResolved inherited resolved::Manager would also result in the
Manager part being removed too early in some cases.  This was also a
poor implementation, so this was switched to a composition model
instead.

Signed-off-by: David Sommerseth <davids@openvpn.net>