client: Fix incorrect empty session_path property
The client would not properly preserve the session path the Session
Manager assigned to the client.  This was a migration error from prior
changes, where a local variable in a method should not be used but the
session_path variable in the main object.

Also ensure this variable is never empty.  This may kick-off some
glib2 errors.

Signed-off-by: David Sommerseth <davids@openvpn.net>