netcfg/dns: Fix glib2 GVariant ref counting issues
The AddNameServers() and AddSearchDomains() methods in the
ResolverSettings class makes use of the glib2::Value::ExtractVector()
which will free the GVariant object containing the input parameters.

This causes an issue when caller of these methods also tries to free the
same value container.  The g_variant_ref_sink() call indicates that
there is another owner holding a reference counter to this object,
which avoids glib2 to avoid complaining about incorrect reference
counting.

Signed-off-by: David Sommerseth <davids@openvpn.net>