configmgr/proxy: Do an extra object existence check in ctor
The OpenVPN3ConfigurationProxy::GetConfig() method could often fail in
the openvpn3-service-client code paths if the configuration manager was
not running.  The reason was that the configuration manager might need
some more time to settle, especially if it is has a lot of persistent
configuration profiles under its control.

By using the DBus::Proxy::Utils::Query::CheckObjectExists() feature in
GDBus++ in the constructor, it can be better checked if the
configuration really exists or not when setting up the proxy.  This
check will also do more attempts with a delay before each check, so the
chances of success increases tremendously.

Signed-off-by: David Sommerseth <davids@openvpn.net>