sessionmgr: Add generic code for sending session and manager signals
This will be used by the new Session Manager which will be based on
GDBus++.

The session manager proxies AttentionRequired and StatusChange signals
from the backend VPN process - which is inaccessible directly by
end-users, only the session manager has access to these processes.

Here the needed framework for subscribing and proxying these signals are
provided.  In addition is the code needed to send the StatusManagerEvent
signals which is sent each time a VPN session is created or destroyed.

Signed-off-by: David Sommerseth <davids@openvpn.net>