netcfg/dns: Fix missing virtual destructor compiler warnings
The NetCfg::DNS::ResolverBackendInterface class declares several
pure virtual functions in an abstract class.  Newer compilers will
complain about the inheriting classes not declaring a destructor
and not having it declared virtual.

It is not expected that the destructor will be used in most
implementations, so we declare a default virtual destructor in the
NetCfg::DNS::ResolverBackendInterface class.

Signed-off-by: David Sommerseth <davids@openvpn.net>