tests: Make PlatformInfo unit-test run without D-Bus
The PlatformInfo.DBus unit test did no longer detect missing D-Bus
connection.  This is due to DBus::Connection::Create() throwing an
exception if it fails; which is a new behaviour with GDBus++.

Catch the DBus::Connection::Exception and skip the test if this happens.

This change will allow the unit-test to more easily run in build systemd
preparing the Debian and RPM packages.

Signed-off-by: David Sommerseth <davids@openvpn.net>