events: Relocate AttentionReq and StatusEvent
The separation between C++ event handling objects and the D-Bus signal
handling of the AttentionRequired and StatusChange signals can be a
bit confusing.

This is an attempt to reorganise the code a bit, moving everything
related to the C++ event handling into src/events.  A C++ namespace
is also introduced, the Events namespace.  In that regards, the
StatusEvent object has been renamed to just Events::Status.

Signed-off-by: David Sommerseth <davids@openvpn.net>