dbus/signals: Refactor AttentionRequired and StatusChange signals
Currently the AttentionRequired and StatusChange signal sending and
proxy handling was put into the Session Manager code scope.  This can be
used more places.  So slightly rearrange the code and the scoping.

Now these signals are located in src/dbus/signals, using the ::Signals
C++ namespace.

The SessionManagerEvent is relocated to the same ::Signals C++, for
consistency.  Since this signal is expected only to be used by the
Session Manager it is kept in the src/sessionmgr source directory.

Signed-off-by: David Sommerseth <davids@openvpn.net>