client: Rework how backend VPN client sends StatusChange/AttentionRequired
The LogSender class is doing too much and it shouldn't care about
StatusChange signals at all.  In addition, the implementation of
AttentionRequired signals was quite low-level.

This change is somewhat similar to the previous change, where signal
handling and sending is handled in the ::Signals::StatusChange
(src/dbus/signals/statuschange.hpp) class.  In addition, there is also a
dedicated class for sending the AttentionRequired signal too, in
::Signals::AttentionRequired (src/dbus/signals/attention-required.hpp).

Signed-off-by: David Sommerseth <davids@openvpn.net>