build: Refactor ENABLE_DEBUG macro in netcfg-dns-direct-file-selftest
The netcfg-dns-direct-file-selftest test program was the only user of
the ENABLE_DEBUG macro, which made it impossible to build it without
this macros explicitly enabled.  That meant the 'debug_internal'
configure setting had to always be enabled.

Since the dependent features in resolvconf-file.hpp was only used by
this test program, this adds a bit refactoring to move that
functionality into the main test program itself.  This way, there are no
hard dependencies on 'debug_internal'/ENABLE_DEBUG any more.

Signed-off-by: David Sommerseth <davids@openvpn.net>