build: Redo the ENABLE_OVPNDCO macro setup
Since OpenVPN 3 Core 3.8.5, it is now possible to do the build using the
ENABLE_OVPNDCO macro from the build-config.h header.  This also makes it
easier to rework the configuration logic related to the 'dco'
configuration setting.

In addition, the dco feature had to be explicitly enabled to be built.
If set to 'auto', it would not be enabled.  Since it should be possible
to build with DCO support in all circumstances (there are no external
dependencies for the build), we can enable this by default when set to
'auto'.

Signed-off-by: David Sommerseth <davids@openvpn.net>