log: Cleanup LogMetaData and LogMetaDataValue
This just cleans up some coding style related details in the LogMetaData
and LogMetaDataValue classes, streamlining it towards the model used in
all the migrated code.  The std::shared_ptr is created using a Create()
method, the constructor is private and so on.  Basically enforce using
smart pointers.

This brings up a challenge with the LogMetaData class, since there are
some code paths depending on the copy constructor.  To provide a simpler
implementation, the LogMetaData::Duplicate() method was added.  This
will return a new LogMetaData object re-using the
shared_ptr<LogMetaDataValue> references to the values itself.  This will
allow the LogMetaData objects to be handled differently (the source and
the duplicated copy) without destroying any of the values itself.  The
LogMetaDataValue elements are constant once created and can only be
deleted.

Signed-off-by: David Sommerseth <davids@openvpn.net>