events: Extend Events::Log() to also carry signal sender details
This extends the Events::Log() constructor parsing a GVariant object to
also take a DBus::Signals::Target object which can contain more details
about the sender of this event.

With this information available, it becomes easier to extract the needed
meta data for log events in the net.openvpn.v3.log service.

Signed-off-by: David Sommerseth <davids@openvpn.net>