GDBus++: Migrate openvpn3-service-configmgr
This provides the complete migration of the
openvpn3-service-configmgr service to use GDBus++.

The build system config has been extended to build and install
the binary, man page, D-Bus policies and D-Bus service config.

Additionally, these changes add support for sending a signal
out when a configuration profile is added or removed.

Signed-off-by: Razvan Cojocaru <razvan.cojocaru@openvpn.com>