sessionmgr: Improve retrieve sessions helper method
The helper_retrieve_sessions() method was implemented in a fragile way
due to the use of the std::static_pointer_cast<Session>() usage.

The DBus::Object::Manager::GetAllObjects() will return all
DBus::Object::Base based objects it manages.  Since this is a base
class, the application implementing these objects can use more derived
classes to represent the D-Bus object it implements.  Even the Session
Manager uses two DBus::Object::Base derived classes: SrvHandler and
Session.

The for-loop in this code skipped the SrvHandler object because of the
path filtering happening first.  Without this filtering, the
std::static_pointer_cast<Session> would result casting the SrvHandler
to a Session object - which would fail badly at runtime.

Instead, use std::dynamic_pointer_cast<Session>.  This ensures that the
returned pointer will be nullptr if the derived object is not a Session
object.  This also helps simplify the code further, as now there is no
need to filter out an object based on the D-Bus object path - the cast
object pointer will be nullptr for all the objects we're not interested
in.

Thanks goes to Răzvan Cojocaru for spotting this defect.

Reported-by: Răzvan Cojocaru <razvan.cojocaru@openvpn.com>
Signed-off-by: David Sommerseth <davids@openvpn.net>