events/log: Extend Events::Log to carry a LogTag::Ptr
A log event can be referring to a specific log sender which are
identified by a LogTag object.  Instead of passing this information
outside of the Events::Log object to the log writer, let this
information travel with the Events::Log.

This has the advantage of ensuring multiple Log events at the same time
will not get mixed up with a different LogTag.

Signed-off-by: David Sommerseth <davids@openvpn.net>