GDBus++/log: Migration of net.openvpn.v3.log / openvpn3-service-log
This is a full migration of the logging service in the OpenVPN 3 Linux
stack.

This migration removes some features which is not relevant any more,
like stand-alone logging for various services.  This depended on
services doing logging via broadcast signals, which is no longer
happening.

The internal structures are also more streamlined and each service
attaching itself to the logger service will now be managed in the
AttachedService() object.  And services allowing Log and StatusChange
signals to be forwarded to end-users will be handled by the
ProxyLogEvents object.  This object is owned by the AttachedService
object, which makes the relation between them more clear.

With this refactoring, the binary is also renamed to
openvpn3-service-log, which matches better the D-Bus service name.

Signed-off-by: David Sommerseth <davids@openvpn.net>