GDBus++/ovpn3cli: Migrate 'openvpn3-admin version'
The previous implementation hid most of the --services feature in
cmd_version() inside some #ifdef macros.  This gives very little
benefit, from a coding perspective.

This has taken the "service version extraction" logic out of the old
cmd_version() function and put it inside the cmd_version_services()
function.  This has been re-implemented using a slightly different
approach to retrieve the version information from each of the services
listed here.

What is new in this approach is that the service executable name is also
extracted as well, by parsing /proc/$PID/cmdline.  This gives a more
exact information on the service itself.

Signed-off-by: David Sommerseth <davids@openvpn.net>