log/journald: Add O3_LOG_SENDER meta data for log service
It can be hard to identify the sender of log events since various
systemd versions on various Linux distributions will have slightly
different variations in the journal meta data.

To avoid that, lets introduce a constant O3_LOG_SENDER meta data which
is the D-Bus service name.  This should be expected to be unique for the
OpenVPN 3 Linux stack.

Signed-off-by: David Sommerseth <davids@openvpn.net>