log/journald: Extend the journald parser to include the updated log service
The openvpn3-service-logger has been renamed, so the current journald filter
will not match on log events sent from the new log service.

During the development process of the new logger, the service was
renamed to 'openvpn3-service-log-dev'.  The matching of the old log
service name has been preserved as well as the temporarily new one.

With the new full release, the log service will now also set the
O3_LOG_SENDER meta data record to 'net.openvpn.v3.log'.  So this will
become the default match filter for the future.  This filter has also
been added and the old ones with the binary names can be removed in a
future release, like v25.

The SYSLOG_IDENTIFIER filter has been kept, as that will mostly be
correct if the log service is incorrectly configured to use the syslog
instead of the journald logging.

This commit will also re-enable the journal-log-parse test program.

Signed-off-by: David Sommerseth <davids@openvpn.net>