sessionmgr/proxy: Add Session::GetConfigName()
Retrieve the configuration profile name for a specific VPN session.
This will contain the configuration profile name at the time of the
creation of the VPN session.  If the configuration profile name has
been changed after the VPN session was started, the old name is
preserved in the return value.

Signed-off-by: David Sommerseth <davids@openvpn.net>