netcfg/dns: Preserve device name in systemd-resolved D-Bus proxy
There are no direct ways to extract the device name of a
NetCfg::DNS::resolved::Link object.  This is needed when sending the
NetworkChange event signals.

This extends the NetCfg::DNS::resolved::Link object to preserve a device
name.  The  NetCfg::DNS::resolved::Manager::RetrieveLink() is updated to
send the device name when creating the Link object.

Signed-off-by: David Sommerseth <davids@openvpn.net>