netcfg: Fix incorrect signal subscription D-Bus API
The net.openvpn.v3.netcfg.NotificationUnsubscribe() argument to the
subscriber argument is not really optional.  The v21 and older
versions used an ugly hack where it didn't expose this variable in
the introspection while it was extracted if present in the C++
callback method.  This may break some D-Bus protocol specifications
and is not possible via GDBus++.  This change will make this field
mandatory.  However, the content can be an empty string to achieve
the behaviour of the old implementation.

The net.openvpn.v3.netcfg.NotificationSubscriberList() was also not
properly declared and did not list any return types at all.

Signed-off-by: David Sommerseth <davids@openvpn.net>