GDBus++/ovpn3cli: Migrate 'openvpn3-admin netcfg-service'
During the migration of this command, several bugs and lacking
features in the net.openvpn.v3.log and net.openvpn.v3.netcfg
services was found.  But with those fixes applied, the migration
of this command was pretty much straight forward.  The bigger
blocks of changes are more simplification of the code, due to
the needed functionality is now available via some APIs.

Signed-off-by: David Sommerseth <davids@openvpn.net>