ovpn3cli/arghelpers: Add optional DBus::Connection to retrieve_config_path()
The GDBus++ DBus::Connection object will disconnect the D-Bus connection
on destruction.  So if a connection is already created, calling this
function will close the connection - because glib2 shares the D-Bus
connection setup internally in the library.

To circumvent this behaviour, allow the caller to pass the
DBus::Connection object it uses.  This will keep the connection open as
it was created in an outer-scope than inside this connection.

Signed-off-by: David Sommerseth <davids@openvpn.net>