GDBus++/ovpn3cli: Migrate 'openvpn3 config-manage'
This needed a bit refactoring as there were several duplicated code
blocks which could be reused in a shared scope.

There was also an issue that the retrieve_config_path() call closed
the DBus::Connection when returning.  This could be solved by passing
the DBus::Connection already available in the calling function.

Also switch to use smart-pointers for the configmgr proxy to avoid
similar issues when passing the proxy object around.

Signed-off-by: David Sommerseth <davids@openvpn.net>