dbus: Improve object ownership GetUID/GetPID call error scenarios
When the CheckACL() and CheckOwnerAccess() methods calls the GetUID()
method, a exception may be throw if the  lookup failed.  Instead of
resulting in an exception throw, handle this as an authorization
failure instead.

This should be fine as the it indicates that the calling process does
no longer exists, so we should not trust this request for access.

Signed-off-by: David Sommerseth <davids@openvpn.net>