sessionmgr: Improve log_forwarders access across multiple threads
The std::map implementation can easily crash a program if one worker
thread is reading the std::map and another one is manipulating it
(adding/removing elements).  A quick fix is to introduce locking on
the access to the log_forwarders std::map object.

All reads and writes are now locked so only a single worker thread can
access this map, thus serializing all the access for it.

In this code path, this is not performance critical so this should be
fine.  Adding and removing log forwarders is not something which is
timing critical at all.

Signed-off-by: David Sommerseth <davids@openvpn.net>