sessionmgr: Session::method_ready() uses wrong exception string for parsing
When looking for the 'net.openvpn.v3.error.ready' error domain in the
error message, it's needed to use the low-level method to retrieve the
detailed error message.  The .GetRawError() only provides the end-user
friendly error message without such details.

And since the caller of the Ready() method may also need to parse this
string, it is needed to pass the string from .what() too when throwing
the ReadyException back.

Signed-off-by: David Sommerseth <davids@openvpn.net>